--- a/api/auth.go
+++ b/api/auth.go
@@ -3,41 +3,30 @@ package api
 import (
 	"encoding/base64"
 	"encoding/json"
-	"errors"
 	"io/ioutil"
 	"os/user"
 	"strings"
 )
 
-func GetRegistryAuth(imageName string) (string, error) {
-	registryURL, username, password, err := GetAuth(imageName)
-	if err != nil {
-		return "", &popagateError{originalErr: err}
-	}
+var AuthInfo = make(map[string]map[string]string)
 
-	stringRegistryAuth := `{"username": "DOCKERUSERNAME", "password": "DOCKERPASSWORD", "email": null, "serveraddress": "DOCKERREGISTRYURL"}`
-
-	stringRegistryAuth = strings.Replace(stringRegistryAuth, "DOCKERUSERNAME", username, 1)
-	stringRegistryAuth = strings.Replace(stringRegistryAuth, "DOCKERPASSWORD", password, 1)
-	stringRegistryAuth = strings.Replace(stringRegistryAuth, "DOCKERREGISTRYURL", registryURL, 1)
-	RegistryAuth := base64.URLEncoding.EncodeToString([]byte(stringRegistryAuth))
-
-	return RegistryAuth, nil
-}
-
-func GetAuth(imageName string) (string, string, string, error) {
+func GetAuth() {
 	usr, err := user.Current()
 	if err != nil {
-		return "", "", "", &popagateError{
-			originalErr: err,
-			newErr:      errors.New(" :unable to find current user")}
+		// unable to find user
+		AuthInfo["quay"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+		AuthInfo["dockerhub"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+		return
 	}
+
 	// TODO: the config can be in many places
 	// try to find them and use them; https://github.com/docker/docker-py/blob/e9fab1432b974ceaa888b371e382dfcf2f6556e4/docker/auth.py#L269
 	dockerAuth, err := ioutil.ReadFile(usr.HomeDir + "/.docker/config.json")
 	if err != nil {
 		// we'll just try accessing the public access docker hubs/quay
-		return "", "", "", nil
+		AuthInfo["quay"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+		AuthInfo["dockerhub"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+		return
 	}
 
 	type AuthData struct {
@@ -46,39 +35,54 @@ func GetAuth(imageName string) (string, string, string, error) {
 	data := &AuthData{}
 	err = json.Unmarshal([]byte(dockerAuth), data)
 	if err != nil {
-		return "", "", "", &popagateError{
-			originalErr: err,
-			newErr:      errors.New(" :unable to unmarshal auth info")}
+		AuthInfo["quay"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+		AuthInfo["dockerhub"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+		return
 	}
 
-	encodedAuth := "placeholder"
-	registryURL := "placeholder"
 	// TODO: we are only checking for dockerHub and quay.io
 	// registries, we should probably be exhaustive in future.
-	if strings.Contains(imageName, "quay") {
-		// quay
-		encodedAuth = data.Auths["quay.io"]["auth"]
-		registryURL = "quay.io"
-	} else {
-		encodedAuth = data.Auths["https://index.docker.io/v1/"]["auth"]
-		registryURL = "https://index.docker.io/v1/"
-	}
+	dockerEncodedAuth := data.Auths["https://index.docker.io/v1/"]["auth"]
+	dockerRegistryURL := "https://index.docker.io/v1/"
+	quayEncodedAuth := data.Auths["quay.io"]["auth"]
+	quayRegistryURL := "quay.io"
 
-	if encodedAuth == "" {
-		return "", "", "", &popagateError{
-			newErr: errors.New(" :unable to find any auth info in ~/.docker/config.json")}
+	if dockerEncodedAuth == "" {
+		AuthInfo["dockerhub"] = map[string]string{"registryURL": "", "username": "", "password": ""}
 	}
-
-	yourAuth, err := base64.StdEncoding.DecodeString(encodedAuth)
+	if quayEncodedAuth == "" {
+		AuthInfo["quay"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+	}
+	dockerAuth, err = base64.StdEncoding.DecodeString(dockerEncodedAuth)
 	if err != nil {
-		return "", "", "", &popagateError{
-			originalErr: err,
-			newErr:      errors.New(" :unable to base64 decode auth info")}
+		AuthInfo["dockerhub"] = map[string]string{"registryURL": "", "username": "", "password": ""}
 	}
-	userPass := FormatRegistryAuth(string(yourAuth))
-	username := userPass[0]
-	password := userPass[1]
+	quayAuth, err := base64.StdEncoding.DecodeString(quayEncodedAuth)
+	if err != nil {
+		AuthInfo["quay"] = map[string]string{"registryURL": "", "username": "", "password": ""}
+	}
 
-	return registryURL, username, password, nil
+	dockerUserPass := FormatRegistryAuth(string(dockerAuth))
+	dockerUsername := dockerUserPass[0]
+	dockerPassword := dockerUserPass[1]
+
+	quayUserPass := FormatRegistryAuth(string(quayAuth))
+	quayUsername := quayUserPass[0]
+	quayPassword := quayUserPass[1]
+
+	dockerStringRegistryAuth := `{"username": "DOCKERUSERNAME", "password": "DOCKERPASSWORD", "email": null, "serveraddress": "DOCKERREGISTRYURL"}`
+	dockerStringRegistryAuth = strings.Replace(dockerStringRegistryAuth, "DOCKERUSERNAME", dockerUsername, 1)
+	dockerStringRegistryAuth = strings.Replace(dockerStringRegistryAuth, "DOCKERPASSWORD", dockerPassword, 1)
+	dockerStringRegistryAuth = strings.Replace(dockerStringRegistryAuth, "DOCKERREGISTRYURL", dockerRegistryURL, 1)
+	dockerRegistryAuth := base64.URLEncoding.EncodeToString([]byte(dockerStringRegistryAuth))
+
+	quayStringRegistryAuth := `{"username": "quayUSERNAME", "password": "quayPASSWORD", "email": null, "serveraddress": "quayREGISTRYURL"}`
+	quayStringRegistryAuth = strings.Replace(quayStringRegistryAuth, "quayUSERNAME", quayUsername, 1)
+	quayStringRegistryAuth = strings.Replace(quayStringRegistryAuth, "quayPASSWORD", quayPassword, 1)
+	quayStringRegistryAuth = strings.Replace(quayStringRegistryAuth, "quayREGISTRYURL", quayRegistryURL, 1)
+	quayRegistryAuth := base64.URLEncoding.EncodeToString([]byte(quayStringRegistryAuth))
+
+	AuthInfo["dockerhub"] = map[string]string{"registryURL": dockerRegistryURL, "username": dockerUsername, "password": dockerPassword, "RegistryAuth": dockerRegistryAuth}
+	AuthInfo["quay"] = map[string]string{"registryURL": quayRegistryURL, "username": quayUsername, "password": quayPassword, "RegistryAuth": quayRegistryAuth}
 
 }