cache auth info instead of hitting disk for every container.

As a result of this change, Pulling docker Images is now about 10 times faster. 
`go test -timeout 3m -race -cover -v -bench=BenchmarkPullDockerImage -run=XXX ./...` 

old: `153,701 ns/op`

new(with normal map): `15,915 ns/op`

newer(with syncMap): 16,228 ns/op,  16,007 ns/op